Refacored Report: removed redundant lines and json file
