--- hedera-state-validator/src/main/java/com/hedera/statevalidation/reporting/InvariantProperty.java
@@ -1,11 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-package com.hedera.statevalidation.reporting;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.FIELD)
-public @interface InvariantProperty {}
